mp4 writer: fix calculation of sample duration in stts box
The code calculating a sample duration for the stts box was in
fact calculating the previous sample duration. This also meant
that the very first sample duration was usually completely
incorrect, although it had the added benefit of keeping the
synchronisation between the audio/video tracks if they did not
happen to start at the same time offset.

Note that we ought to also implement writting an edit list box
to correctly maintain the sync between audio / video.