mp4_writer: fix timestamp overlow
The mp4 writer does not cope with initial timestamps bigger than
2^31us. Since it internally uses 32 bits signed integers to store
the DTS difference between consecutive samples, this will end up
overflowing the 32bits integer.
If the overflow happens to generate a big negative value, this will
in turn result in an mp4 file with with all the frames having a
timestamp of 0us.

To avoid this problem, store the initial timestamp separately as a
64bits integer to avoid any potential overflow.